Rename Screenshot 2025-07-24 210238.png to step3_cropped_numberPlate.png
This is the cropped number plate image which is later sent for ocr to detect the text. The second model localizes and creates a bounding box around the vehicle number plate from the cropped vehicle input. It then crops around the bounding box and then sends the number plate to the OCR Model